Refactor converters
